requirements.txt
pyspark==3.5.0
delta-spark==3.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
aif360==0.6.0
fairlearn==0.10.0
diffprivlib==0.6.3
causalnex==0.12.1
dowhy==0.11.1
ray==2.8.0
sentence-transformers==2.2.2
spacy==3.7.2
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
pyyaml==6.0.1
requests==2.31.0
streamlit==1.28.0
plotly==5.17.0
pytest==7.4.3
pytest-cov==4.1.0
matplotlib==3.8.0
seaborn==0.13.0
